pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "readmission-predictor-mimic"
version = "0.1.0"
description = "Temporal knowledge graph for hospital readmission prediction using MIMIC-IV"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = "MIT"
dependencies = [
    "duckdb>=1.1",
    "rdflib>=7.0",
    "pandas>=2.0",
    "numpy",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "pydantic>=2.0",
    "pydantic-settings",
    "google-cloud-bigquery[pandas]>=3.10",
    "db-dtypes>=1.0",
    "requests>=2.28",
    "oxrdflib>=0.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "hypothesis>=6.0",
]
graph = [
    "neo4j>=5.0",
    "networkx>=3.0",
]
gnn = [
    "torch>=2.1",
    "torch_geometric>=2.5",
    "torch-sparse>=0.6",
    "torch-scatter>=2.1",
    "transformers>=4.35",
]
cloud = [
    "google-cloud-storage>=2.14",
    "google-cloud-secret-manager>=2.18",
]
conversational = [
    "anthropic>=0.39",
    "streamlit>=1.40",
    "plotly>=5.0",
    # Phase D/E/F (v2 MCP integration + deterministic SQL validator):
    # mcp Python SDK for stdio + HTTP MCP clients; sqlglot for SQL static
    # analysis; fastmcp for the bq-validator MCP server. All three are
    # required for the bq-validator path; mcp + sqlglot are required at
    # client time. fastmcp is only needed when running the local stdio
    # server.
    "mcp>=1.0",
    "sqlglot>=25.0",
    "fastmcp>=0.2",
]
# Phase 8 causal-inference pipeline. Deps are lazy-imported by the
# estimator + survival modules; the base install stays slim. Install
# with `uv pip install --python .venv/bin/python -e .[causal]` when
# running the Phase 8d+ estimators or 8h/8i diagnostics.
causal = [
    "econml>=0.14",
    "lifelines>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "bigquery: marks tests that require gcloud auth and BigQuery access (run with '-m bigquery')",
    "bigquery_e2e: full BigQuery loader test, ~$0.35 per run (run with '-m bigquery_e2e')",
    "live_llm: tests that hit the real Anthropic API (skip by default; run with '-m live_llm')",
]
addopts = "-v -m 'not bigquery_e2e and not live_llm'"
